requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.10
paho-mqtt==2.1.0
pycryptodome==3.21.0
requests==2.32.3
tuya-connector-python==0.1.2
tuya-iot-py-sdk==0.6.6
urllib3==2.2.3
websocket-client==1.8.0
